# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive anonymous form builder and response collection system",
        "title": "AnoQ Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "AnoQ Support",
            "email": "support@anoq.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/form": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of forms created by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "List user's forms",
                "responses": {
                    "200": {
                        "description": "List of forms",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "forms": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/model.Form"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new form with title, description, and slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Create a new form",
                "parameters": [
                    {
                        "description": "Form creation data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateFormRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Form created successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "form": {
                                    "$ref": "#/definitions/model.Form"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Form with this slug already exists",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/form/slug/{slug}": {
            "get": {
                "description": "Get a form by its slug identifier (public endpoint)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Get form by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form details",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "form": {
                                    "$ref": "#/definitions/model.Form"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Form not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/form/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing form's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Update a form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Form update data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateFormRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "form": {
                                    "$ref": "#/definitions/model.Form"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or form ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied: you don't own this form",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Form not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete an existing form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forms"
                ],
                "summary": "Delete a form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Form deleted successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid form ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Form not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CreateFormRequest": {
            "description": "Request payload for creating a new form",
            "type": "object",
            "required": [
                "slug",
                "title"
            ],
            "properties": {
                "description": {
                    "description": "Form description",
                    "type": "string",
                    "example": "A form to collect customer feedback"
                },
                "questions": {
                    "description": "Optional list of questions to create with the form",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CreateQuestionRequest"
                    }
                },
                "slug": {
                    "description": "URL-friendly identifier (required)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "customer-feedback-2023"
                },
                "title": {
                    "description": "Form title (required)",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Customer Feedback Form"
                }
            }
        },
...
        "model.Form": {
            "description": "Form structure containing all form details",
            "type": "object",
            "required": [
                "slug",
                "title"
            ],
            "properties": {
                "author": {
                    "description": "Form author details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.User"
                        }
                    ]
                },
                "author_id": {
                    "description": "Form creator's user ID",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "created_at": {
                    "description": "Form creation timestamp",
                    "type": "string",
                    "example": "2023-01-01T10:00:00Z"
                },
                "description": {
                    "description": "Form description",
                    "type": "string",
                    "example": "A form to collect customer feedback"
                },
                "id": {
                    "description": "Form unique identifier",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440002"
                },
                "modified_at": {
                    "description": "Last modification timestamp",
                    "type": "string",
                    "example": "2023-01-01T10:00:00Z"
                },
                "questions": {
                    "description": "List of questions in the form",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Question"
                    }
                },
                "slug": {
                    "description": "URL-friendly form identifier",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "customer-feedback-2023"
                },
                "status": {
                    "description": "Form status (open/closed)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.FormStatus"
                        }
                    ],
                    "example": "open"
                },
                "title": {
                    "description": "Form title",
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "Customer Feedback Form"
                }
            }
        },
...
        "model.FormStatus": {
            "type": "string",
            "enum": [
                "open",
                "closed"
            ],
            "x-enum-varnames": [
                "FormStatusOpen",
                "FormStatusClosed"
            ]
        },
...
        "model.UpdateFormRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UpdateQuestionRequest"
                    }
                },
                "status": {
                    "$ref": "#/definitions/model.FormStatus"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "model.User": {
            "description": "User account information",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Account creation timestamp",
                    "type": "string",
                    "example": "2023-01-01T10:00:00Z"
                },
                "email": {
                    "description": "User email address",
                    "type": "string",
                    "example": "user@example.com"
                },
                "family_name": {
                    "description": "User's family name",
                    "type": "string",
                    "example": "Doe"
                },
                "given_name": {
                    "description": "User's given name",
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "description": "User unique identifier",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2023-01-01T10:00:00Z"
                },
                "username": {
                    "description": "User's chosen username",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionToken": {
            "description": "Session token for authentication",
            "type": "apiKey",
            "name": "X-Session-Token",
            "in": "header"
        }
    }
}